Refactor InnerMonologue, InnerMonologueSchema, PlanSchema
